+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## Version 1.1.0
+
+- Add support for `html`, `markdown` and `vue` files
+- Fix detection of `standard` installation
+
 ## Version 1.0.1
 
 Update URL to screenshot in Readme